+++ hdl/common/include/base/dummy.hpp
